+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [Unreleased]
+- added phone number conversion to international format for Lithuania
+
 ## [1.3.0]
 - **Breaking change**: changed how Package::isOffloadPostcodeRequired() function works, now it expects to be given Package object, with main service, channel and servicePackage (if applied)
 - if offloadPostcode is set on receiver Address object, when getting receiver address for registration it will return array with offloadPostcode format